refac client services, syncronizers and fetcher for beacon sync
