+++ packages/client/lib/config.ts
@@ -130,6 +130,13 @@ export interface ConfigOptions {
   maxPerRequest?: number
 
   /**
+   * Max jobs to be enqued in a fetcher at any given time
+   *
+   * Default: `100``
+   */
+  maxFetcherJobs?: number
+
+  /**
    * Number of peers needed before syncing
    *
    * Default: `1`
@@ -220,6 +227,7 @@ export class Config {
   public static readonly TRANSPORTS_DEFAULT = ['rlpx', 'libp2p']
   public static readonly PORT_DEFAULT = 30303
   public static readonly MAXPERREQUEST_DEFAULT = 50
+  public static readonly MAXFETCHERJOBS_DEFAULT = 100
   public static readonly MINPEERS_DEFAULT = 1
   public static readonly MAXPEERS_DEFAULT = 25
   public static readonly DNSADDR_DEFAULT = '8.8.8.8'
@@ -240,6 +248,7 @@ export class Config {
   public readonly saveReceipts: boolean
   public readonly txLookupLimit: number
   public readonly maxPerRequest: number
+  public readonly maxFetcherJobs: number
   public readonly minPeers: number
   public readonly maxPeers: number
   public readonly dnsAddr: string
@@ -277,6 +286,7 @@ export class Config {
     this.saveReceipts = options.saveReceipts ?? false
     this.txLookupLimit = options.txLookupLimit ?? 2350000
     this.maxPerRequest = options.maxPerRequest ?? Config.MAXPERREQUEST_DEFAULT
+    this.maxFetcherJobs = options.maxFetcherJobs ?? Config.MAXFETCHERJOBS_DEFAULT
     this.minPeers = options.minPeers ?? Config.MINPEERS_DEFAULT
     this.maxPeers = options.maxPeers ?? Config.MAXPEERS_DEFAULT
     this.dnsAddr = options.dnsAddr ?? Config.DNSADDR_DEFAULT
