+++ packages/client/lib/sync/fetcher/fetcher.ts
@@ -57,9 +57,9 @@ export abstract class Fetcher<JobTask, JobResult, StorageItem> extends Readable 
   protected processed: number // number of processed tasks, awaiting the write job
   protected finished: number // number of tasks which are both processed and also finished writing
   protected running: boolean
-  protected errored?: Error
   protected reading: boolean
   protected destroyWhenDone: boolean // Destroy the fetcher once we are finished processing each task.
+  errored?: Error
 
   private _readableState?: {
     // This property is inherited from Readable. We only need `length`.
@@ -137,6 +137,8 @@ export abstract class Fetcher<JobTask, JobResult, StorageItem> extends Readable 
   tasks(): JobTask[] {
     return []
   }
+
+  nextTasks(): void {}
 
   /**
    * Enqueue job
@@ -292,6 +294,7 @@ export abstract class Fetcher<JobTask, JobResult, StorageItem> extends Readable 
    * Process next task
    */
   next() {
+    this.nextTasks()
     const job = this.in.peek()
     if (!job) {
       this.debug(`No job found on next task, skip next job execution.`)
@@ -439,11 +442,7 @@ export abstract class Fetcher<JobTask, JobResult, StorageItem> extends Readable 
     }
     this.write()
     this.running = true
-    const tasks = this.tasks()
-    for (const task of tasks) {
-      this.enqueueTask(task)
-    }
-    this.debug(`Enqueued num=${tasks.length} tasks`)
+
     while (this.running) {
       if (!this.next()) {
         if (this.finished === this.total && this.destroyWhenDone) {
